Added more regions servers which are started by regions supervisor

+++ makefile
@@ -1,7 +1,8 @@
 all:
-	rm private/game_log.log
+	@[ -f private/game_log.log ] && rm private/game_log.log || echo "Log file not exists"
 	rebar compile
 	rebar -f generate
 
 doc:
+	set -v off
 	rebar doc
